archive/artifact_backend/requirements.txt
Django
python-dotenv
mysqlclient
djangorestframework
django-cors-headers
SPARQLWrapper
django-rest-swagger
drf-yasg
aiohttp
